refactor: blind=blindsearch | informed=localsearch
